requirements.txt
office365-rest-python-client
llama_parse
python-dotenv
langchain
llama-index


